pyproject.toml
[build-system]
requires = ["hatchling>=1.21"]
build-backend = "hatchling.build"

[project]
name = "career-hub"
version = "0.1.0"
description = "AI-first centralized professional life database and job application CLI"
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [
    { name = "Luis Fernando Molina", email = "fernando@siavato.com" },
]
keywords = ["job-search", "career", "ai", "claude", "cli", "applications"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business",
    "Topic :: Utilities",
]
dependencies = [
    "click>=8.0",
    "rich>=13.0",
    "python-docx>=1.0",
    "pdfplumber>=0.10",
    "PyYAML>=6.0",
    "tomli>=2.0; python_version < '3.11'",
]

[project.optional-dependencies]
scan = ["apify-client>=1.0", "requests>=2.31"]
web = ["flask>=3.0"]
dev = ["pytest>=8.0", "ruff>=0.4", "pytest-cov>=5.0"]
all = ["apify-client>=1.0", "requests>=2.31", "flask>=3.0"]

[project.scripts]
career-hub = "jobsearch.__main__:cli"

[project.urls]
Homepage = "https://github.com/fernandosiavato/career-hub"
Repository = "https://github.com/fernandosiavato/career-hub"
Issues = "https://github.com/fernandosiavato/career-hub/issues"
Changelog = "https://github.com/fernandosiavato/career-hub/blob/main/CHANGELOG.md"

[tool.hatch.build.targets.wheel]
packages = ["src/jobsearch"]

[tool.hatch.build.targets.wheel.force-include]
"src/jobsearch/templates" = "jobsearch/templates"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP"]
ignore = [
    "E501",  # line too long — formatter responsibility
    "E402",  # module import not at top — lazy imports for optional deps
    "E741",  # ambiguous variable name (e.g. ``l``) — pre-existing usage
    "B904",  # raise-without-from-inside-except — pre-existing, low impact
    "F401",  # unused-import — preserved for explicit re-exports
    "F841",  # unused-variable — pre-existing, low impact
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-ra"

[tool.coverage.run]
source = ["src/jobsearch"]
omit = [
    # Modules below need fixtures/mocks beyond the Phase 5 scope and are
    # therefore excluded from the package-level coverage gate. They are
    # tested through manual smoke runs and downstream CLI integration tests.
    "src/jobsearch/__main__.py",      # CLI dispatch (Click commands)
    "src/jobsearch/apply.py",         # /apply orchestrator
    "src/jobsearch/cv_builder.py",    # DOCX file copy
    "src/jobsearch/scanner.py",       # ATS scraping
    "src/jobsearch/scanner_apify.py", # Apify integration
    "src/jobsearch/web.py",           # Flask web UI
]

[tool.coverage.report]
fail_under = 60
show_missing = true
skip_covered = false
